requirements.txt
asgiref==3.2.5
Django==3.0.4
django-cors-headers==3.4.0
djangorestframework==3.11.1
pytz==2019.3
sqlparse==0.3.1
